requirements.txt
Django
djangorestframework
djangorestframework-simplejwt[crypto]
django-cors-headers
psycopg2-binary
python-decouple
celery
redis
pillow
requests
stripe
python-dotenv
dj-rest-auth
django-allauth
django-filter
drf-spectacular
whitenoise
gunicorn
fal-client
boto3
django-storages